+++ tests/PhpSpreadsheetTests/Custom/Complex.php
@@ -6,106 +6,168 @@ use PhpOffice\PhpSpreadsheet\Exception;
 
 class Complex
 {
-    private $realPart = 0;
+    /**
+     * @constant    Euler's Number.
+     */
+    const EULER = 2.7182818284590452353602874713526624977572;
 
-    private $imaginaryPart = 0;
+    /**
+     * @constant    Regexp to split an input string into real and imaginary components and suffix
+     */
+    const NUMBER_SPLIT_REGEXP =
+        '` ^
+            (                                   # Real part
+                [-+]?(\d+\.?\d*|\d*\.?\d+)          # Real value (integer or float)
+                ([Ee][-+]?[0-2]?\d{1,3})?           # Optional real exponent for scientific format
+            )
+            (                                   # Imaginary part
+                [-+]?(\d+\.?\d*|\d*\.?\d+)          # Imaginary value (integer or float)
+                ([Ee][-+]?[0-2]?\d{1,3})?           # Optional imaginary exponent for scientific format
+            )?
+            (                                   # Imaginary part is optional
+                ([-+]?)                             # Imaginary (implicit 1 or -1) only
+                ([ij]?)                             # Imaginary i or j - depending on whether mathematical or engineering
+            )
+        $`uix';
 
-    private $suffix;
+    /**
+     * @var    float    $realPart    The value of of this complex number on the real plane.
+     */
+    protected $realPart = 0.0;
 
-    public static function _parseComplex($complexNumber)
+    /**
+     * @var    float    $imaginaryPart    The value of of this complex number on the imaginary plane.
+     */
+    protected $imaginaryPart = 0.0;
+
+    /**
+     * @var    string    $suffix    The suffix for this complex number (i or j).
+     */
+    protected $suffix;
+
+
+    /**
+     * Validates whether the argument is a valid complex number, converting scalar or array values if possible
+     *
+     * @param     mixed    $complexNumber   The value to parse
+     * @return    array
+     * @throws    Exception    If the argument isn't a Complex number or cannot be converted to one
+     */
+    private static function parseComplex($complexNumber)
     {
-        //    Test for real number, with no imaginary part
+        // Test for real number, with no imaginary part
         if (is_numeric($complexNumber)) {
             return [$complexNumber, 0, null];
         }
 
-        //    Fix silly human errors
-        if (strpos($complexNumber, '+-') !== false) {
-            $complexNumber = str_replace('+-', '-', $complexNumber);
-        }
-        if (strpos($complexNumber, '++') !== false) {
-            $complexNumber = str_replace('++', '+', $complexNumber);
-        }
-        if (strpos($complexNumber, '--') !== false) {
-            $complexNumber = str_replace('--', '-', $complexNumber);
-        }
+        // Fix silly human errors
+        $complexNumber = str_replace(
+            ['+-', '-+', '++', '--'],
+            ['-', '-', '+', '+'],
+            $complexNumber
+        );
 
-        //    Basic validation of string, to parse out real and imaginary parts, and any suffix
-        $validComplex = preg_match('/^([\-\+]?(\d+\.?\d*|\d*\.?\d+)([Ee][\-\+]?[0-2]?\d{1,3})?)([\-\+]?(\d+\.?\d*|\d*\.?\d+)([Ee][\-\+]?[0-2]?\d{1,3})?)?(([\-\+]?)([ij]?))$/ui', $complexNumber, $complexParts);
+        // Basic validation of string, to parse out real and imaginary parts, and any suffix
+        $validComplex = preg_match(
+            self::NUMBER_SPLIT_REGEXP,
+            $complexNumber,
+            $complexParts
+        );
 
         if (!$validComplex) {
-            //    Neither real nor imaginary part, so test to see if we actually have a suffix
+            // Neither real nor imaginary part, so test to see if we actually have a suffix
             $validComplex = preg_match('/^([\-\+]?)([ij])$/ui', $complexNumber, $complexParts);
             if (!$validComplex) {
-                throw new Exception('COMPLEX: Invalid complex number');
+                throw new Exception('Invalid complex number');
             }
-            //    We have a suffix, so set the real to 0, the imaginary to either 1 or -1 (as defined by the sign)
+            // We have a suffix, so set the real to 0, the imaginary to either 1 or -1 (as defined by the sign)
             $imaginary = 1;
             if ($complexParts[1] === '-') {
                 $imaginary = 0 - $imaginary;
             }
-
             return [0, $imaginary, $complexParts[2]];
         }
 
-        //    If we don't have an imaginary part, identify whether it should be +1 or -1...
+        // If we don't have an imaginary part, identify whether it should be +1 or -1...
         if (($complexParts[4] === '') && ($complexParts[9] !== '')) {
             if ($complexParts[7] !== $complexParts[9]) {
                 $complexParts[4] = 1;
                 if ($complexParts[8] === '-') {
                     $complexParts[4] = -1;
                 }
-                //    ... or if we have only the real and no imaginary part (in which case our real should be the imaginary)
             } else {
+                // ... or if we have only the real and no imaginary part
+                //  (in which case our real should be the imaginary)
                 $complexParts[4] = $complexParts[1];
                 $complexParts[1] = 0;
             }
         }
 
-        //    Return real and imaginary parts and suffix as an array, and set a default suffix if user input lazily
-        return [$complexParts[1], $complexParts[4], !empty($complexParts[9]) ? $complexParts[9] : 'i'];
+        // Return real and imaginary parts and suffix as an array, and set a default suffix if user input lazily
+        return [
+            $complexParts[1],
+            $complexParts[4],
+            !empty($complexParts[9]) ? $complexParts[9] : 'i'
+        ];
     }
 
-    //    function _parseComplex()
 
-    public function __construct($realPart, $imaginaryPart = null, $suffix = 'i')
+    public function __construct($realPart = 0.0, $imaginaryPart = null, $suffix = 'i')
     {
         if ($imaginaryPart === null) {
             if (is_array($realPart)) {
-                //    We have an array of (potentially) real and imaginary parts, and any suffix
-                list($realPart, $imaginaryPart, $suffix) = array_values($realPart) + [0.0, 0.0, 'i'];
+                // We have an array of (potentially) real and imaginary parts, and any suffix
+                list ($realPart, $imaginaryPart, $suffix) = array_values($realPart) + [0.0, 0.0, 'i'];
+                if ($suffix === null) {
+                    $suffix = 'i';
+                }
             } elseif ((is_string($realPart)) || (is_numeric($realPart))) {
-                //    We've been given a string to parse to extract the real and imaginary parts, and any suffix
-                list($realPart, $imaginaryPart, $suffix) = self::_parseComplex($realPart);
+                // We've been given a string to parse to extract the real and imaginary parts, and any suffix
+                list($realPart, $imaginaryPart, $suffix) = self::parseComplex($realPart);
             }
         }
 
-        //    Set parsed values in our properties
+        // Set parsed values in our properties
         $this->realPart = (float) $realPart;
         $this->imaginaryPart = (float) $imaginaryPart;
         $this->suffix = strtolower($suffix);
     }
 
+    /**
+     * Gets the real part of this complex number
+     *
+     * @return Float
+     */
     public function getReal()
     {
         return $this->realPart;
     }
 
+    /**
+     * Gets the imaginary part of this complex number
+     *
+     * @return Float
+     */
     public function getImaginary()
     {
         return $this->imaginaryPart;
     }
 
+    /**
+     * Gets the suffix of this complex number
+     *
+     * @return String
+     */
     public function getSuffix()
     {
         return $this->suffix;
     }
 
-    public function __toString()
+    public function format()
     {
-        $str = '';
+        $str = "";
         if ($this->imaginaryPart != 0.0) {
-            if (abs($this->imaginaryPart) != 1.0) {
+            if (\abs($this->imaginaryPart) != 1.0) {
                 $str .= $this->imaginaryPart . $this->suffix;
             } else {
                 $str .= (($this->imaginaryPart < 0.0) ? '-' : '') . $this->suffix;
@@ -113,14 +175,19 @@ class Complex
         }
         if ($this->realPart != 0.0) {
             if (($str) && ($this->imaginaryPart > 0.0)) {
-                $str = '+' . $str;
+                $str = "+" . $str;
             }
             $str = $this->realPart . $str;
         }
         if (!$str) {
-            $str = '0.0';
+            $str = "0.0";
         }
 
         return $str;
+    }
+
+    public function __toString()
+    {
+        return $this->format();
     }
 }